chg: 调整索引数据结构

+++ internal/serializers/cache.go
@@ -1,12 +1,12 @@
 package serializers
 
 type DataCache struct {
-	Category string `json:"category"`
-	Language string `json:"language"`
-	Name     string `json:"name"`
-	Code     string `json:"code" `
-	Type     int8   `json:"type"`
-	Number   uint32 `json:"number"`
-	Desc     string `json:"desc"`
-	Weight   int32  `json:"weight"`
+	Category string  `json:"category"`
+	Language string  `json:"language"`
+	Name     string  `json:"name"`
+	Code     string  `json:"code" `
+	Type     int8    `json:"type"`
+	Number   uint32  `json:"number"`
+	Desc     string  `json:"desc"`
+	Weight   float32 `json:"weight"`
 }
